refactoring course details files
