+++ src/styles/CourseDetailsStyle.css
@@ -1,0 +1,81 @@
+.course-details {
+    display: grid;
+    grid-template-rows: auto auto auto;
+    gap: 1rem;
+    width: 100%;
+    padding: 2rem;
+}
+
+.course-info,
+.course-sections,
+.course-comments {
+    background-color: #f7f7f71c;
+    border-radius: 0.5rem;
+    padding: 1rem;
+    color: white;
+}
+
+.course-info {
+    display: grid;
+    grid-template-columns: 1fr 2fr;
+    gap: 1rem;
+}
+
+.course-info img {
+    width: 1000px;
+    height: 350px;
+    border-radius: 8px;
+}
+
+.coursedetail {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    margin-left: 15px;
+}
+
+.course-image {
+    width: 100%;
+    max-width: 600px;
+    height: auto;
+    margin-bottom: 1rem;
+}
+
+.comment-input {
+    display: flex;
+    align-items: center;
+}
+
+.comment-input input {
+    padding: 0.5rem;
+    border: none;
+    border-radius: 0.5rem;
+    color: white;
+}
+
+.send-icon,
+.back-button {
+    color: white;
+    cursor: pointer;
+    margin-left: 10px;
+    font-size: 1.5rem;
+    background-color: #007BFF;
+    border-radius: 50%;
+    padding: 0.5rem;
+}
+
+.send-icon:hover,
+.back-button:hover {
+    background-color: white;
+    color: #007BFF;
+}
+
+.loading {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100%;
+    margin-top: 100px;
+    font-size: 1.3rem;
+    color: rgba(255, 255, 255, 0.449);
+}
